[installer]: separate out server ide to separate package
This is to ensure the correct team ownership
