+++ installer/pkg/components/server/ide/OWNERS
@@ -1,0 +1,9 @@
+
+options:
+  no_parent_owners: true
+
+approvers:
+  - engineering-ide
+
+labels:
+  - "team: IDE"
