+++ installer/pkg/components/server/ide/constants.go
@@ -1,0 +1,13 @@
+// Copyright (c) 2021 Gitpod GmbH. All rights reserved.
+// Licensed under the GNU Affero General Public License (AGPL).
+// See License-AGPL.txt in the project root for license information.
+
+package ide
+
+import (
+	"github.com/gitpod-io/gitpod/installer/pkg/common"
+)
+
+const (
+	Component = common.ServerComponent
+)
